requirements.txt
twitchio>=2.7.0
aiohttp>=3.8.0
websockets==15.0.1
python-dotenv>=1.0.0

# PyTorch 2.3, CPU-версия (совместима с TTS)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu

TTS==0.22.0
sounddevice==0.4.6
# TTS и torchaudio требуют numpy 1.x
numpy<2.0
redis==6.4.0
transliterate==1.10.2

silero-tts==0.0.5
